Poll for pending HVAC states [fixes #21]
When setting mode=AUTO while the car is sleeping, the target state
might not be reached until more than 30 seconds later (the delay
we use after sending an HVAC command). In that case, what happens
is:

1. Set hvac_mode=AUTO -> send hvac command -> write state
2. After 30 seconds, poll TeslaFi
3. If the state has not been updated yet, this will write state=OFF
4. Some time later (3 min default), TeslaFi has the updated state
   and will report state=AUTO

The transition from 2->3->4 causes the climate entity to flip
states unexpectedly.

For `lock` and `alarm_control_panel` platforms, we retain a pending
or target state, and on the next refresh if the pending state has
not been met yet, we schedule another refresh, until the target
state is met. This adds the same technique for climate mode.

We only do this for mode, because it's the one that is most likely
to be set from a sleeping state.

In addition, we'll also ask TeslaFi to add the `wake=` parameter
which will cause the API to wake the car first and wait a few
seconds for the car to wake up before sending the real command.

Long term it would be great to have a better (and more reusable)
way of doing this. Ideally the HA service call should not return
until the target state is fully verified, but given the polling
nature of both the integration and TeslaFi itself, this could
cause problems in HA.