Add start/stop charge switch
This switch can take the place of the binary_sensor.charging entity,
where the switch can be toggled off/on in order to stop/start charging.

The command(s) must be enabled in TeslaFi for this to work, and the
way the entity appears in Home Assistant might be less desirable
for some - that is why the entity is disabled by default.

If you want to keep the binary_sensor, but also still have the
ability to stop/start charging, you can set the switch entity to
hidden, and use some other means to toggle the switch, like with
a Button or automation calling switch.* services.