Use longer wake allowance
The TeslaFi API supports adding a `wake=N` argument that instructs
TeslaFi to wait for up to `N` seconds for the car to wake up before
sending the requested command. In my testing, wake=10 was sufficient
to allow the car to wake up, but that testing was performed at home
on wifi.

Issues #39 and #47 report that 10 seconds may not be enough time to
wake the car fully, but 30 seconds appears to be more reliable.

If the car is already awake when a command is being sent, then the
argument will have no effect. So this should be a safe change.